+++ activerecord/lib/active_record/migration.rb
@@ -131,6 +131,16 @@ module ActiveRecord
     end
   end
 
+  class InvalidMigrationTimestampError < MigrationError # :nodoc:
+    def initialize(version = nil, name = nil)
+      if version && name
+        super("Invalid timestamp #{version} for migration file: #{name}.\nTimestamp should be in form YYYYMMDDHHMMSS.")
+      else
+        super("Invalid timestamp for migration. Timestamp should be in form YYYYMMDDHHMMSS.")
+      end
+    end
+  end
+
   class PendingMigrationError < MigrationError # :nodoc:
     include ActiveSupport::ActionableError
 
@@ -493,7 +503,11 @@ module ActiveRecord
   #
   #    20080717013526_your_migration_name.rb
   #
-  # The prefix is a generation timestamp (in UTC).
+  # The prefix is a generation timestamp (in UTC). Timestamps should not be
+  # modified manually. To validate that migration timestamps adhere to the
+  # format Rails expects, you can use the following configuration option:
+  #
+  #    config.active_record.validate_migration_timestamps = true
   #
   # If you'd prefer to use numeric prefixes, you can turn timestamped migrations
   # off by setting:
@@ -1201,6 +1215,8 @@ module ActiveRecord
   class MigrationContext
     attr_reader :migrations_paths, :schema_migration, :internal_metadata
 
+    MIGRATION_TIMESTAMP_REGEX = /\A(\d{4})(\d{2})(\d{2})(\d{2})(\d{2})(\d{2})\z/ # :nodoc:
+
     def initialize(migrations_paths, schema_migration = nil, internal_metadata = nil)
       if schema_migration == SchemaMigration
         ActiveRecord.deprecator.warn(<<-MSG.squish)
@@ -1316,6 +1332,9 @@ module ActiveRecord
       migrations = migration_files.map do |file|
         version, name, scope = parse_migration_filename(file)
         raise IllegalMigrationNameError.new(file) unless version
+        if validate_timestamp?
+          raise InvalidMigrationTimestampError.new(version, name) unless valid_migration_timestamp?(version)
+        end
         version = version.to_i
         name = name.camelize
 
@@ -1331,6 +1350,9 @@ module ActiveRecord
       file_list = migration_files.filter_map do |file|
         version, name, scope = parse_migration_filename(file)
         raise IllegalMigrationNameError.new(file) unless version
+        if validate_timestamp?
+          raise InvalidMigrationTimestampError.new(version, name) unless valid_migration_timestamp?(version)
+        end
         version = schema_migration.normalize_migration_number(version)
         status = db_list.delete(version) ? "up" : "down"
         [status, version, (name + scope).humanize]
@@ -1373,6 +1395,24 @@ module ActiveRecord
 
       def parse_migration_filename(filename)
         File.basename(filename).scan(Migration::MigrationFilenameRegexp).first
+      end
+
+      def validate_timestamp?
+        ActiveRecord.timestamped_migrations && ActiveRecord.validate_migration_timestamps
+      end
+
+      def valid_migration_timestamp?(version)
+        match_data = version.match(MIGRATION_TIMESTAMP_REGEX)
+        return false unless match_data
+
+        year, month, day, hour, minute, second = match_data.captures
+        begin
+          Time.new(year, month, day, hour, minute, second)
+        rescue ArgumentError
+          return false
+        end
+
+        true
       end
 
       def move(direction, steps)