+++ activerecord/CHANGELOG.md
@@ -1,3 +1,10 @@
+*   Add `active_record.config.validate_migration_timestamps` option for validating migration timestamps.
+
+    When set, validates that the timestamp prefix for a migration is in the form YYYYMMDDHHMMSS.
+    This is designed to prevent migration timestamps from being modified by hand.
+
+    *Adrianna Chang*
+
 *   In cases where MySQL returns `warning_count` greater than zero, but returns no warnings when
     the `SHOW WARNINGS` query is executed, `ActiveRecord.db_warnings_action` proc will still be
     called with a generic warning message rather than silently ignoring the warning(s).